+++ AdvancedTeamCreation.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30804.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdvancedTeamCreation", "AdvancedTeamCreation\AdvancedTeamCreation.csproj", "{1731B4D5-1F9E-4680-A86B-BE85B648117A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A936C1D4-9853-4882-B00D-45BDD1DC4F37}
+	EndGlobalSection
+EndGlobal
