+++ AdvancedTeamCreation/API/Enums/Enums.cs
@@ -1,0 +1,15 @@
+﻿namespace AdvancedTeamCreation.API
+{
+    public enum SpawnLocation
+    {
+        Normal,
+        Escape,
+        SCP106,
+        SurfaceNuke,
+        SCP012,
+        SCP079,
+        SCP096,
+        SCP173,
+        Shelter
+    }
+}
