+++ AdvancedTeamCreation/API/Features/Points.cs
@@ -1,0 +1,11 @@
+﻿namespace AdvancedTeamCreation.API
+{
+    using UnityEngine;
+
+    public class RoomPoint
+    {
+        public SpawnLocation Type;
+        public Vector3 Position;
+        public Vector2 Direction;
+    }
+}
