--- TeamsEXILED.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30621.155
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeamsEXILED", "TeamsEXILED\TeamsEXILED.csproj", "{1731B4D5-1F9E-4680-A86B-BE85B648117A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1731B4D5-1F9E-4680-A86B-BE85B648117A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {10DF3B64-52E2-4B0A-B45E-D056C75309EF}
-	EndGlobalSection
-EndGlobal
