--- TeamsEXILED/API/Enums/Enums.cs
@@ -1,15 +1,0 @@
-﻿namespace TeamsEXILED.API
-{
-    public enum SpawnLocation
-    {
-        Normal,
-        Escape,
-        SCP106,
-        SurfaceNuke,
-        SCP012,
-        SCP079,
-        SCP096,
-        SCP173,
-        Shelter
-    }
-}
