--- TeamsEXILED/API/Features/Points.cs
@@ -1,11 +1,0 @@
-﻿using UnityEngine;
-
-namespace TeamsEXILED.API
-{
-    public class RoomPoint
-    {
-        public SpawnLocation Type;
-        public Vector3 Position;
-        public Vector2 Direction;
-    }
-}
