--- TeamsEXILED/Patches/RefNextTeam.cs
@@ -1,25 +1,0 @@
-﻿using Exiled.API.Features;
-using HarmonyLib;
-using System;
-
-namespace TeamsEXILED.Patches
-{
-    [HarmonyPatch(typeof(Respawning.RespawnTickets), nameof(Respawning.RespawnTickets.DrawRandomTeam))]
-    public class RefNextTeam
-    {
-        public static void Postfix(ref Respawning.SpawnableTeamType __result)
-        {
-            try
-            {
-                if (!MainPlugin.Singleton.EventHandlers.HasReference)
-                {
-                    TeamMethods.RefNextTeamSpawn(__result);
-                }
-            }
-            catch (Exception e)
-            {
-                Log.Error(e);
-            }
-        }
-    }
-}
